refactor: TodoPage 컴포넌트 패턴 정리 (HomePage와 통일)
